+++ lib/src/models/models.dart
@@ -4,3 +4,4 @@ export 'user/student.dart';
 export 'user/user.dart';
 export 'catalogue/rule.dart';
 export 'catalogue/country.dart';
+export 'academic/archive.dart';