+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 ## 0.0.1-beta2
 > Release date: 08/Oct/2021
 * Add catalogue section
+* Add academic section
 * Rename `UMMobileAPI` to `UMMobileSDK`
 * Rename `getUser()` to `getInformation()`